+++ lib/internal/http2/core.js
@@ -110,13 +110,11 @@ const {
   onStreamRead,
   kAfterAsyncWrite,
   kMaybeDestroy,
-  kUpdateTimer
+  kUpdateTimer,
+  kSession,
+  setStreamTimeout
 } = require('internal/stream_base_commons');
-const {
-  kTimeout,
-  setUnrefTimeout,
-  validateTimerDuration
-} = require('internal/timers');
+const { kTimeout } = require('internal/timers');
 const { isArrayBufferView } = require('internal/util/types');
 
 const { FileHandle } = internalBinding('fs');
@@ -163,7 +161,6 @@ const kSelectPadding = Symbol('select-padding');
 const kSentHeaders = Symbol('sent-headers');
 const kSentTrailers = Symbol('sent-trailers');
 const kServer = Symbol('server');
-const kSession = Symbol('session');
 const kState = Symbol('state');
 const kType = Symbol('type');
 const kWriteGeneric = Symbol('write-generic');
@@ -2546,35 +2543,7 @@ const setTimeout = {
   configurable: true,
   enumerable: true,
   writable: true,
-  value: function(msecs, callback) {
-    if (this.destroyed)
-      return;
-
-    // Type checking identical to timers.enroll()
-    msecs = validateTimerDuration(msecs);
-
-    // Attempt to clear an existing timer lear in both cases -
-    //  even if it will be rescheduled we don't want to leak an existing timer.
-    clearTimeout(this[kTimeout]);
-
-    if (msecs === 0) {
-      if (callback !== undefined) {
-        if (typeof callback !== 'function')
-          throw new ERR_INVALID_CALLBACK();
-        this.removeListener('timeout', callback);
-      }
-    } else {
-      this[kTimeout] = setUnrefTimeout(this._onTimeout.bind(this), msecs);
-      if (this[kSession]) this[kSession][kUpdateTimer]();
-
-      if (callback !== undefined) {
-        if (typeof callback !== 'function')
-          throw new ERR_INVALID_CALLBACK();
-        this.once('timeout', callback);
-      }
-    }
-    return this;
-  }
+  value: setStreamTimeout
 };
 Object.defineProperty(Http2Stream.prototype, 'setTimeout', setTimeout);
 Object.defineProperty(Http2Session.prototype, 'setTimeout', setTimeout);
