+++ lib/internal/stream_base_commons.js
@@ -11,12 +11,23 @@ const {
   streamBaseState
 } = internalBinding('stream_wrap');
 const { UV_EOF } = internalBinding('uv');
-const { errnoException } = require('internal/errors');
+const {
+  codes: {
+    ERR_INVALID_CALLBACK
+  },
+  errnoException
+} = require('internal/errors');
 const { owner_symbol } = require('internal/async_hooks').symbols;
+const {
+  kTimeout,
+  setUnrefTimeout,
+  validateTimerDuration
+} = require('internal/timers');
 
 const kMaybeDestroy = Symbol('kMaybeDestroy');
 const kUpdateTimer = Symbol('kUpdateTimer');
 const kAfterAsyncWrite = Symbol('kAfterAsyncWrite');
+const kSession = Symbol('session');
 
 function handleWriteReq(req, data, encoding) {
   const { handle } = req;
@@ -178,6 +189,38 @@ function onStreamRead(arrayBuffer) {
   }
 }
 
+function setStreamTimeout(msecs, callback) {
+  if (this.destroyed)
+    return;
+
+  this.timeout = msecs;
+
+  // Type checking identical to timers.enroll()
+  msecs = validateTimerDuration(msecs);
+
+  // Attempt to clear an existing timer in both cases -
+  //  even if it will be rescheduled we don't want to leak an existing timer.
+  clearTimeout(this[kTimeout]);
+
+  if (msecs === 0) {
+    if (callback !== undefined) {
+      if (typeof callback !== 'function')
+        throw new ERR_INVALID_CALLBACK();
+      this.removeListener('timeout', callback);
+    }
+  } else {
+    this[kTimeout] = setUnrefTimeout(this._onTimeout.bind(this), msecs);
+    if (this[kSession]) this[kSession][kUpdateTimer]();
+
+    if (callback !== undefined) {
+      if (typeof callback !== 'function')
+        throw new ERR_INVALID_CALLBACK();
+      this.once('timeout', callback);
+    }
+  }
+  return this;
+}
+
 module.exports = {
   createWriteWrap,
   writevGeneric,
@@ -186,4 +229,6 @@ module.exports = {
   kAfterAsyncWrite,
   kMaybeDestroy,
   kUpdateTimer,
+  kSession,
+  setStreamTimeout
 };
